Move P2pBasicTestTimeGetter to p2p-test-utils
