+++ dns_server/src/crawler_p2p/crawler_manager/tests/mock_manager.rs
@@ -39,9 +39,9 @@ use p2p::{
         ConnectivityService, NetworkingService, SyncingEventReceiver,
     },
     protocol::NETWORK_PROTOCOL_CURRENT,
-    testing_utils::P2pBasicTestTimeGetter,
     types::peer_id::PeerId,
 };
+use p2p_test_utils::P2pBasicTestTimeGetter;
 use tokio::sync::mpsc;
 
 use crate::{
